refactor: create package for each module #27
