+++ app/src/main/java/pl/gawor/android/tayckner/common/util/SharedPrefManager.kt
@@ -1,9 +1,9 @@
-package pl.gawor.android.tayckner.util
+package pl.gawor.android.tayckner.common.util
 
 import android.content.Context
 import android.util.Log
 import androidx.fragment.app.Fragment
-import pl.gawor.android.tayckner.model.CredentialsModel
+import pl.gawor.android.tayckner.habit_tracker.model.CredentialsModel
 
 object SharedPrefManager {
