+++ app/src/main/java/pl/gawor/android/tayckner/habit_tracker/model/Habit.kt
@@ -1,4 +1,4 @@
-package pl.gawor.android.tayckner.model
+package pl.gawor.android.tayckner.habit_tracker.model
 
 data class Habit(
     var id: Long,
