+++ app/src/main/java/pl/gawor/android/tayckner/habit_tracker/model/ResponseModel.kt
@@ -1,4 +1,4 @@
-package pl.gawor.android.tayckner.model
+package pl.gawor.android.tayckner.habit_tracker.model
 
 data class ResponseModel<T>(
     val code: String,
