+++ app/src/main/java/pl/gawor/android/tayckner/habit_tracker/service/HabitApi.kt
@@ -1,7 +1,7 @@
-package pl.gawor.android.tayckner.service
+package pl.gawor.android.tayckner.habit_tracker.service
 
-import pl.gawor.android.tayckner.model.Habit
-import pl.gawor.android.tayckner.model.ResponseModel
+import pl.gawor.android.tayckner.habit_tracker.model.Habit
+import pl.gawor.android.tayckner.habit_tracker.model.ResponseModel
 import retrofit2.Response
 import retrofit2.http.*
