+++ app/src/main/java/pl/gawor/android/tayckner/habit_tracker/service/RetrofitInstance.kt
@@ -1,4 +1,4 @@
-package pl.gawor.android.tayckner.service
+package pl.gawor.android.tayckner.habit_tracker.service
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
