+++ app/src/main/res/layout/fragment_habit_tracker.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@drawable/background_main"
-    tools:context=".HabitTrackerFragment">
+    tools:context=".habit_tracker.fragment.HabitTrackerFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout_header"
