+++ app/src/main/res/layout/fragment_my_habits.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@drawable/background_main"
-    tools:context=".MyHabitsFragment">
+    tools:context=".habit_tracker.fragment.MyHabitsFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout_header"
